Bunch of updates

+++ assets/src/scripts/services/dittyService.js
@@ -1,11 +1,19 @@
 import { getRenderedItems } from "./httpService";
 import { displayTypeExists } from "../utils/helpers";
 
+/**
+ * @deprecated since 4.0 - Use v4 preview system with actions.refreshPreview()
+ * Initialize Ditty display (v3 legacy)
+ */
 export const initializeDitty = (dittyEl, displayType, args) => {
   if (!displayTypeExists(dittyEl, displayType)) return false;
   jQuery(dittyEl)[`ditty_${displayType}`](args);
 };
 
+/**
+ * @deprecated since 4.0 - Use v4 preview system with actions.refreshPreview()
+ * Update Ditty display template (v3 legacy)
+ */
 export const updateDittyDisplayTemplate = (dittyEl, display) => {
   if (!displayTypeExists(dittyEl, display.type)) return false;
   const prevType = dittyEl.dataset.type;
@@ -23,6 +31,10 @@ export const updateDittyDisplayTemplate = (dittyEl, display) => {
   dittyEl.dataset.type = display.type;
 };
 
+/**
+ * @deprecated since 4.0 - Use v4 preview system with actions.refreshPreview()
+ * Update Ditty display type (v3 legacy)
+ */
 export const updateDittyDisplayType = (dittyEl, displayType) => {
   if (!displayTypeExists(dittyEl, displayType)) return false;
   const prevType = dittyEl.dataset.type;
@@ -38,7 +50,12 @@ export const updateDittyDisplayType = (dittyEl, displayType) => {
   }
 };
 
+/**
+ * @deprecated since 4.0 - Use v4 preview system with actions.refreshPreview()
+ * Update display options (v3 legacy)
+ */
 export const updateDisplayOptions = (dittyEl, option, value) => {
+  if (!dittyEl) return false;
   const displayType = dittyEl.dataset.type;
   if (!displayTypeExists(dittyEl, displayType)) return false;
   dittyEl[`_ditty_${displayType}`].options(option, value);
@@ -63,7 +80,12 @@ export const getDisplayItems = async (items, layouts, returnData) => {
   }
 };
 
+/**
+ * @deprecated since 4.0 - Use v4 preview system with actions.refreshPreview()
+ * Add display items to Ditty (v3 legacy)
+ */
 export const addDisplayItems = (dittyEl, displayItems) => {
+  if (!dittyEl) return false;
   const displayType = dittyEl.dataset.type;
   if (!displayTypeExists(dittyEl, displayType)) return false;
   displayItems.map((displayItem) =>
@@ -72,40 +94,38 @@ export const addDisplayItems = (dittyEl, displayItems) => {
 };
 
 /**
- * Delete an item from the Ditty display
+ * @deprecated since 4.0 - Use v4 preview system with actions.refreshPreview()
+ * Delete an item from the Ditty display (v3 legacy)
  * @param {element} dittyEl
  * @param {object} item
  */
 export const deleteDisplayItems = (dittyEl, item) => {
+  if (!dittyEl) return false;
   const displayType = dittyEl.dataset.type;
   if (!displayTypeExists(dittyEl, displayType)) return false;
   dittyEl[`_ditty_${displayType}`].deleteItem(item.item_id);
 };
 
 /**
- * Update the Ditty display items
- * @param {element} dittyEl
- * @param {object} items
- */
-// export const updateDisplayItems = (dittyEl, displayItems) => {
-//   const displayType = dittyEl.dataset.type;
-//   if (!displayTypeExists(dittyEl, displayType)) return false;
-//   dittyEl["_ditty_" + displayType].loadItems(displayItems, "update");
-// };
-
-/**
- * Update the Ditty display items
+ * @deprecated since 4.0 - Use v4 preview system with actions.refreshPreview()
+ * Replace the Ditty display items (v3 legacy)
  * @param {element} dittyEl
  * @param {object} items
  */
 export const replaceDisplayItems = (dittyEl, displayItems) => {
+  if (!dittyEl) return false;
   const displayType = dittyEl.dataset.type;
   if (!displayTypeExists(dittyEl, displayType)) return false;
   dittyEl[`_ditty_${displayType}`].loadItems(displayItems, "static");
 };
 
+/**
+ * @deprecated since 4.0 - Use v4 preview system
+ * Show a specific item (v3 legacy)
+ */
 export const showItem = (item) => {
   const dittyEl = document.getElementById("ditty-editor__ditty");
+  if (!dittyEl) return false;
   const displayType = dittyEl.dataset.type;
   dittyEl[`_ditty_${displayType}`].showItem(item.item_id);
 };
